+++ test/webdav_test.dart
@@ -1,88 +1,15 @@
 import 'dart:convert';
 import 'dart:io';
 
-import 'package:browser_launcher/browser_launcher.dart';
 import 'package:nextcloud/nextcloud.dart';
 import 'package:test/test.dart';
 
-class Config {
-  const Config({
-    this.host,
-    this.username,
-    this.password,
-    this.shareUser,
-    this.testDir,
-    this.email,
-    this.storageLocation,
-  });
+import 'config.dart';
 
-  factory Config.fromJson(Map<String, dynamic> json) => Config(
-        host: Uri.parse(json['host']),
-        username: json['username'],
-        password: json['password'],
-        shareUser: json['shareUser'],
-        // normalised path (remove trailing slash)
-        testDir: json['testDir'].endsWith('/')
-            ? json['testDir'].substring(0, json['testDir'].length - 1)
-            : json['testDir'],
-        email: json['email'],
-        storageLocation: json['storageLocation'],
-      );
-
-  final Uri host;
-  final String username;
-  final String password;
-  final String shareUser;
-  final String testDir;
-  final String email;
-  final String storageLocation;
-}
-
-// TODO: add more tests
 void main() {
-  final config =
-      Config.fromJson(json.decode(File('config.json').readAsStringSync()));
-  final client = NextCloudClient.withCredentials(
-      config.host, config.username, config.password);
-  group('Nextcloud connection', () {
-    test('Different host urls', () {
-      final urls = [
-        [
-          Uri.parse('http://cloud.test.com/'),
-          'http://cloud.test.com',
-        ],
-        [
-          Uri.parse('https://cloud.test.com:80/'),
-          'https://cloud.test.com:80',
-        ],
-        [
-          Uri(host: 'cloud.test.com'),
-          'https://cloud.test.com',
-        ],
-        [
-          Uri(host: 'cloud.test.com', port: 90),
-          'https://cloud.test.com:90',
-        ],
-        [
-          Uri(host: 'test.com', path: 'cloud'),
-          'https://test.com/cloud',
-        ],
-        [
-          Uri.parse('http://localhost:8081/nextcloud'),
-          'http://localhost:8081/nextcloud',
-        ],
-      ];
+  final config = getConfig();
+  final client = getClient(config);
 
-      for (final url in urls) {
-        final client = NextCloudClient.withCredentials(
-          url[0],
-          config.username,
-          config.password,
-        );
-        expect(client.baseUrl, equals(url[1]));
-      }
-    });
-  });
   group('WebDav', () {
     test('Get status', () async {
       final status = await client.webDav.status();
@@ -318,241 +245,6 @@ void main() {
           'le-custom-prop-value2');
       expect(file.getOtherProp('custom', customNamespaces.keys.elementAt(1)),
           'test-custom-prop-value');
-    });
-  });
-  group('Talk', () {
-    test('Signaling', () async {
-      expect(
-        await client.talk.signalingManagement.getSettings(),
-        isNotNull,
-      );
-    });
-    test('Get conversations', () async {
-      expect(await client.talk.conversationManagement.getUserConversations(),
-          isNotEmpty);
-    });
-    String token;
-    test('Create conversation', () async {
-      token = await client.talk.conversationManagement
-          .createConversation(ConversationType.group, name: 'Test-group');
-      expect(
-        token,
-        isNotNull,
-      );
-    });
-    test('Get conversation', () async {
-      final conversation =
-          await client.talk.conversationManagement.getConversation(token);
-      expect(conversation, isNotNull);
-      expect(conversation.displayName, isNotNull);
-    });
-    test('Rename conversation', () async {
-      expect(
-          await client.talk.conversationManagement.renameConversation(
-            token,
-            'Test-group-2',
-          ),
-          isNull);
-    });
-    test('Allow guests', () async {
-      expect(
-          await client.talk.conversationManagement.allowGuests(token), isNull);
-    });
-    test('Disallow guests', () async {
-      expect(await client.talk.conversationManagement.disallowGuests(token),
-          isNull);
-    });
-    test('Read-only', () async {
-      expect(
-        await client.talk.conversationManagement.setReadOnly(
-          token,
-          ReadOnlyState.readOnly,
-        ),
-        isNull,
-      );
-    });
-    test('Read and write', () async {
-      expect(
-        await client.talk.conversationManagement.setReadOnly(
-          token,
-          ReadOnlyState.readWrite,
-        ),
-        isNull,
-      );
-    });
-    test('Set favorite', () async {
-      expect(
-        await client.talk.conversationManagement
-            .setConversationAsFavorite(token),
-        isNull,
-      );
-    });
-    test('Delete favorite', () async {
-      expect(
-        await client.talk.conversationManagement
-            .deleteConversationAsFavorite(token),
-        isNull,
-      );
-    });
-    test('Set notifications', () async {
-      expect(
-        await client.talk.conversationManagement.setNotificationLevel(
-          token,
-          NotificationLevel.always,
-        ),
-        isNull,
-      );
-    });
-    test('Get participants', () async {
-      expect(
-        (await client.talk.conversationManagement.getParticipants(token))
-            .first
-            .userId,
-        equals(config.username),
-      );
-    });
-    test('Add participant', () async {
-      expect(
-        await client.talk.conversationManagement
-            .addParticipant(token, config.shareUser),
-        isNull,
-      );
-    });
-    test('Set moderator', () async {
-      expect(
-        await client.talk.conversationManagement
-            .promoteUserToModerator(token, config.shareUser),
-        isNull,
-      );
-      expect(
-        await client.talk.conversationManagement
-            .demoteUserFromModerator(token, config.shareUser),
-        isNull,
-      );
-    });
-    test('Get messages', () async {
-      expect(
-        await client.talk.messageManagement.getMessages(token),
-        isNotNull,
-      );
-    });
-    Message message;
-    test('Send message', () async {
-      message = await client.talk.messageManagement.sendMessage(
-        token,
-        'Test',
-      );
-      expect(
-        message,
-        isNotNull,
-      );
-    });
-    test('Reply to message', () async {
-      expect(
-        await client.talk.messageManagement.sendMessage(
-          token,
-          'Test answere',
-          replyTo: message.id,
-        ),
-        isNotNull,
-      );
-    });
-    test('Mention user', () async {
-      expect(
-        await client.talk.messageManagement.getMentionSuggestions(
-          message.token,
-          config.shareUser.substring(0, 2),
-        ),
-        isNotEmpty,
-      );
-    });
-    test('Delete participant', () async {
-      expect(
-        await client.talk.conversationManagement
-            .deleteParticipant(token, config.shareUser),
-        isNull,
-      );
-    });
-    test('Delete conversation', () async {
-      expect(await client.talk.conversationManagement.deleteConversation(token),
-          isNull);
-    });
-  });
-  group('Preview', () {
-    final fullImagePath = '${config.testDir}/preview.png';
-    final rootWithoutUser =
-        config.testDir.split('/files/${config.username}')[1];
-    final imageRootPath = '$rootWithoutUser/preview.png';
-    WebDavFile previewFile;
-
-    setUpAll(() async {
-      await client.webDav
-          .upload(File('test/files/test.png').readAsBytesSync(), fullImagePath);
-      previewFile = await client.webDav.getProps(fullImagePath);
-    });
-
-    test('Get preview by path', () async {
-      expect(await client.preview.getPreviewByPath(imageRootPath, 64, 64),
-          isNotNull);
-    });
-    test('Get preview by id', () async {
-      expect(await client.preview.getPreviewById(previewFile.id, 64, 64),
-          isNotNull);
-    });
-    test('Get preview stream by path', () async {
-      expect(await client.preview.getPreviewStreamByPath(imageRootPath, 64, 64),
-          isNotNull);
-    });
-    test('Get preview stream by id', () async {
-      expect(await client.preview.getPreviewStreamById(previewFile.id, 64, 64),
-          isNotNull);
-    });
-    test('Get thumbnail', () async {
-      expect(
-          await client.preview.getThumbnail(imageRootPath, 64, 64), isNotNull);
-    });
-    test('Get thumbnail stream', () async {
-      expect(await client.preview.getThumbnailStream(imageRootPath, 64, 64),
-          isNotNull);
-    });
-  });
-  group('User', () {
-    test('Get user data', () async {
-      final userdata = await client.user.getUser();
-      expect(userdata.id, equals(config.username));
-      expect(userdata.displayName, equals(config.username));
-      expect(userdata.email, equals(config.email));
-      expect(userdata.storageLocation, equals(config.storageLocation));
-    });
-  });
-  group('Login', () {
-    test('Login flow works', () async {
-      final config =
-          Config.fromJson(json.decode(File('config.json').readAsStringSync()));
-      var client = NextCloudClient.withoutLogin(config.host);
-      final init = await client.login.initLoginFlow();
-      // Linux users might need to create a link: https://github.com/dart-lang/browser_launcher/issues/16
-      await Chrome.start([init.login]);
-      LoginFlowResult _result;
-      while (_result == null) {
-        try {
-          _result = await client.login.pollLogin(init);
-          client = NextCloudClient.withAppPassword(
-            config.host,
-            _result.appPassword,
-          );
-          try {
-            await client.webDav.ls(config.testDir);
-          } catch (e, stacktrace) {
-            print(e);
-            print(stacktrace);
-            fail('Could not read from server after connection!');
-          }
-          // ignore: empty_catches, avoid_catches_without_on_clauses
-        } catch (e) {
-          await Future.delayed(Duration(milliseconds: 500));
-        }
-      }
     });
   });
 }
